Still working on htmx
Got rid of dropdwon of crews and just made it a list of buttons. My goal was to click a crew and it displays members in the member box.  Running into problems that I can't figure out, waiting to hear back from Dr. Baliga.  Pushing what I have for now so we have something that displays at least